requirements.txt
playwright>=1.37.0
playwright-stealth>=1.0.6
fake-useragent>=1.3.0
js2py>=0.74
beautifulsoup4>=4.12.0
selenium>=4.11.0
httpx>=0.24.0
requests>=2.31.0
pypdf2>=3.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
celery>=5.3.0
pandas>=2.1.0
typer>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
python-jose>=3.3.0
customtkinter>=5.2.0
pyperclip>=1.8.0
openpyxl>=3.1.0
sqlmap>=1.7.0
networkx>=3.0
matplotlib>=3.7.0
aiohttp>=3.8.0
dnspython>=2.3.0
pyyaml>=6.0.0
